chore: Renamd rest.py to cms.py
