+++ server/__init__.py
@@ -241,7 +241,7 @@ import server.views.media_picker
 import server.views.sources.search
 import server.views.metadata
 import server.views.platforms
-import server.views.cms.rest
+import server.views.cms.cms
 if (server_app == SERVER_APP_SOURCES) or is_dev_mode():
     import server.views.sources.collection
     import server.views.sources.collectionedit
